Updated Error handling to dbe more informative
Prior to this, uncaught exceptions were not specifically differentiated
from other Errors. Unspecified Errors and Exceptions couldn't be
determinded by the Response of the Bot.
It was also not possible to differentiate between parsing Errors and
a wrong argument count based on the error message the bot relayed to
the user.

This commit updates the messages to include fixes to the above issues.