Changed the Required Permission for some commands
Due to a change in role structure on the N1GP Discord Administrator is
no longer tied to the Moderators role. Moderators should still be able
to use various commands tied to that Permission.

The required permission was changed from "Administrator" to
"ViewAuditLog", because it a permission deeply tied to moderator duty.